Package updates and namespace move (#883)
* Package updates

* Moved IResourceHookExecutorFacade outside Internal sub-namespace, so that uses don't need to reference Internal namespaces when inheriting from resource service.
